+++ src/main/java/com/dna/everythingisbad/block/machines/BlockDryerMachine.java
@@ -6,11 +6,14 @@ import com.dna.everythingisbad.tile.TileDeviceBase;
 import com.dna.everythingisbad.tile.TileDryerMachine;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.init.Items;
+import net.minecraft.init.SoundEvents;
 import net.minecraft.inventory.EntityEquipmentSlot;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fluids.FluidTank;
@@ -36,19 +39,24 @@ public class BlockDryerMachine extends BlockMachineBase {
         if (worldIn.isRemote) {
             return true;
         }
-        TileDeviceBase te = (TileDeviceBase) worldIn.getTileEntity(pos);
-        if (!(te instanceof TileDryerMachine)) {
+        TileDeviceBase tileDeviceBase = (TileDeviceBase) worldIn.getTileEntity(pos);
+        if (!(tileDeviceBase instanceof TileDryerMachine)) {
             return false;
         }
 
         if(playerIn.getItemStackFromSlot(EntityEquipmentSlot.MAINHAND).getItem().equals(Items.BUCKET)){
-            FluidTank fluidTank = te.getFluidHandler().getFluidTank();
+            FluidTank fluidTank = tileDeviceBase.getFluidHandler().getFluidTank();
             if(fluidTank.getFluidAmount() >= 1000){
 
 
                 playerIn.getItemStackFromSlot(EntityEquipmentSlot.MAINHAND).shrink(1);
                 playerIn.inventory.addItemStackToInventory(FluidUtil.getFilledBucket(fluidTank.getFluid()));
-                te.getFluidHandler().drain(1000,true);
+                if(playerIn instanceof EntityPlayerMP){
+                    EntityPlayerMP playerMP = (EntityPlayerMP)playerIn;
+                    playerMP.getEntityWorld().playSound((EntityPlayer)null, playerIn.posX,playerIn.posY,playerIn.posZ,SoundEvents.ITEM_BOTTLE_EMPTY,SoundCategory.PLAYERS,1f,1f);
+
+                }
+                tileDeviceBase.getFluidHandler().drain(1000,true);
             }
 
         }else{